arithmetic_parser

Struct LocatedSpan

pub struct LocatedSpan<Span, T = ()> {
    pub extra: T,
    /* private fields */
}

Code span together with information related to where it is located in the code.

This type is similar to one from the nom_locate crate, but it has slightly different functionality. In particular, this type provides no method to access other parts of the code (which is performed in nom_locateโs LocatedSpan::get_column() among other methods). As such, this allows to safely replace span info without worrying about undefined behavior.

Fields
extra: T

Extra information that can be embedded by the user.

Implementations

impl<Span, T> LocatedSpan<Span, T>

pub fn location_offset(&self) -> usize

The offset represents the position of the fragment relatively to the input of the parser. It starts at offset 0.

pub fn location_line(&self) -> u32

The line number of the fragment relatively to the input of the parser. It starts at line 1.

pub fn get_column(&self) -> usize

The column of the fragment start.

pub fn fragment(&self) -> &Span

The fragment that is spanned. The fragment represents a part of the input of the parser.

pub fn map_extra<U>(self, map_fn: impl FnOnce(T) -> U) -> LocatedSpan<Span, U>

Maps the extra field of this span using the provided closure.

pub fn map_fragment<U>( self, map_fn: impl FnOnce(Span) -> U, ) -> LocatedSpan<U, T>

Maps the fragment field of this span using the provided closure.

impl<Span: Copy, T> LocatedSpan<Span, T>

pub fn as_ref(&self) -> LocatedSpan<Span, &T>

Returns a copy of this span with borrowed extra field.

pub fn copy_with_extra<U>(&self, value: U) -> LocatedSpan<Span, U>

Copies this span with the provided extra field.

pub fn with_no_extra(&self) -> LocatedSpan<Span>

Removes extra field from this span.

impl<'a> LocatedSpan<&'a str>

pub fn from_str<R>(code: &'a str, range: R) -> Self
where InputSpan<'a>: Input, R: RangeBounds<usize>,

Creates a span from a range in the provided code. This is mostly useful for testing.

impl LocatedSpan<usize>

pub fn from_str<'a, R>(code: &'a str, range: R) -> Self
where InputSpan<'a>: Input, R: RangeBounds<usize>,

Creates a location from a range in the provided code. This is mostly useful for testing.

impl<T> LocatedSpan<usize, T>

pub fn to_string(&self, default_name: &str) -> String

Returns a string representation of this location in the form {default_name} at {line}:{column}.

pub fn span<'a>(&self, code: &'a str) -> &'a str

Returns this location in the provided code. It is callerโs responsibility to ensure that this is called with the original code that produced this location.