Type Alias arithmetic_parser::SpannedExpr

pub type SpannedExpr<'a, T> = Spanned<'a, Expr<'a, T>>;

Expr with the associated type and code span.

Aliased Type

struct SpannedExpr<'a, T> {
    pub extra: Expr<'a, T>,
    /* private fields */
}

Fields
extra: Expr<'a, T>

Extra information that can be embedded by the user.

Implementations

impl<Span, T> LocatedSpan<Span, T>

pub fn location_offset(&self) -> usize

The offset represents the position of the fragment relatively to the input of the parser. It starts at offset 0.

pub fn location_line(&self) -> u32

The line number of the fragment relatively to the input of the parser. It starts at line 1.

pub fn get_column(&self) -> usize

The column of the fragment start.

pub fn fragment(&self) -> &Span

The fragment that is spanned. The fragment represents a part of the input of the parser.

pub fn map_extra<U>(self, map_fn: impl FnOnce(T) -> U) -> LocatedSpan<Span, U>

Maps the extra field of this span using the provided closure.

pub fn map_fragment<U>( self, map_fn: impl FnOnce(Span) -> U ) -> LocatedSpan<U, T>

Maps the fragment field of this span using the provided closure.

impl<Span: Copy, T> LocatedSpan<Span, T>

pub fn as_ref(&self) -> LocatedSpan<Span, &T>

Returns a copy of this span with borrowed extra field.

pub fn copy_with_extra<U>(&self, value: U) -> LocatedSpan<Span, U>

Copies this span with the provided extra field.

pub fn with_no_extra(&self) -> LocatedSpan<Span>

Removes extra field from this span.